add pagination for all module

+++ src/order/order.service.ts
@@ -1,26 +1,34 @@
+import { Stock } from './../stock/entities/stock.entity';
+import { OrderDetail } from './entities/orderDetail.entity';
 import { ProductService } from './../product/product.service';
 import { StockService } from './../stock/stock.service';
-import { IResponse } from 'src/utils/interfaces/response.interface';
-import { Repository, Like } from 'typeorm';
+import { Repository, Like, DataSource } from 'typeorm';
 import { Order } from './entities/order.entity';
-import { HttpStatus, Inject, Injectable } from '@nestjs/common';
+import { HttpStatus, Inject, Injectable, Logger } from '@nestjs/common';
 import { CreateOrderDto, FindOrderDto, UpdateOrderDto } from './order.dto';
+import { IResponse, IPaginate } from 'src/interface/response.interface';
+import { paginateResponse } from 'src/utils/hellper';
 
 @Injectable()
 export class OrderService {
   constructor(
     @Inject('ORDER_REPOSITORY')
     private readonly repository: Repository<Order>,
+    @Inject('ORDER_DETAIL_REPOSITORY')
+    private readonly orderDetail: Repository<OrderDetail>,
     private readonly stock: StockService,
     private readonly product: ProductService,
+    @Inject('STOCK_REPOSITORY')
+    private readonly stockRepository: Repository<Stock>,
+    @Inject('DATA_SOURCE') private readonly connection: DataSource,
   ) {}
 
-  async findAll(payload: FindOrderDto): Promise<IResponse> {
+  async findAll(payload: FindOrderDto): Promise<IResponse | IPaginate> {
     try {
       const { offset, limit, withDeleted, search, orderBy, order } = payload;
-      const orders = await this.repository.find({
+      const orders = await this.repository.findAndCount({
         ...(limit && { take: limit }),
-        ...(offset && { skip: offset }),
+        ...(offset && { skip: (offset - 1) * limit }),
         ...(withDeleted === 'true' ? { withDeleted: true } : {}),
         ...(search && {
           where: [
@@ -39,8 +47,7 @@ export class OrderService {
         }),
         ...(orderBy && { order: { [orderBy]: order } }),
       });
-
-      return { data: orders, error: null, status: HttpStatus.OK };
+      return paginateResponse(orders, offset, limit, null, HttpStatus.OK);
     } catch (error) {
       return {
         message: 'Unable to get orders',
@@ -51,6 +58,8 @@ export class OrderService {
   }
 
   async create(payload: CreateOrderDto): Promise<IResponse> {
+    const queryRunner = await this.connection.createQueryRunner();
+    await queryRunner.startTransaction();
     try {
       const data = await this.repository.find({
         order: { invNumber: 'DESC' },
@@ -77,17 +86,21 @@ export class OrderService {
           this.stock.decrement(detail.productId, quantity);
         });
       }
+      await queryRunner.commitTransaction();
       return {
         message: 'Create order successfully',
         error: null,
         status: HttpStatus.OK,
       };
     } catch (error) {
+      await queryRunner.rollbackTransaction();
       return {
         message: 'Unable to create order',
         error: error.message,
         status: HttpStatus.INTERNAL_SERVER_ERROR,
       };
+    } finally {
+      await queryRunner.release();
     }
   }
 
@@ -112,6 +125,8 @@ export class OrderService {
   }
 
   async update(id: string, payload: UpdateOrderDto): Promise<IResponse> {
+    const queryRunner = this.connection.createQueryRunner();
+    await queryRunner.startTransaction();
     try {
       const order = await this.repository.findOneBy({ id });
       if (!order) {
@@ -121,18 +136,50 @@ export class OrderService {
           status: HttpStatus.NOT_FOUND,
         };
       }
-      await this.repository.update(id, payload);
+      order.orderDetails.map(async (detail) => {
+        const productValue = await this.product.findValueProductByUnit(
+          detail.productId,
+          detail.unitId,
+        );
+        const stock = await this.stock.findOne(detail.productId);
+        if (stock) {
+          const quantity = productValue.value * detail.quantity;
+          stock.quantity += quantity;
+          await this.stockRepository.save(stock);
+          Logger.log(`Increase stock success ${quantity}`);
+        }
+      });
+      const updateOrder = await this.repository.preload({ ...payload, id: id });
+      await this.repository.save(updateOrder);
+      payload.orderDetails.map(async (detail) => {
+        const productValue = await this.product.findValueProductByUnit(
+          detail.productId,
+          detail.unitId,
+        );
+        const stock = await this.stock.findOne(detail.productId);
+        if (stock) {
+          const quantity = productValue.value * detail.quantity;
+          stock.quantity -= quantity;
+          await this.stockRepository.save(stock);
+          Logger.log(`Decrease stock success ${quantity}`);
+        }
+      });
+
+      await queryRunner.commitTransaction();
       return {
         message: 'Update order successfully',
         error: null,
         status: HttpStatus.OK,
       };
     } catch (error) {
+      await queryRunner.rollbackTransaction();
       return {
         message: 'Unable to update order',
         error: error.message,
         status: HttpStatus.INTERNAL_SERVER_ERROR,
       };
+    } finally {
+      await queryRunner.release();
     }
   }
 
@@ -146,7 +193,19 @@ export class OrderService {
           status: HttpStatus.NOT_FOUND,
         };
       }
-      await this.repository.delete({ id });
+      if (order) {
+        order.orderDetails.map(async (detail) => {
+          const productValue = await this.product.findValueProductByUnit(
+            detail.productId,
+            detail.unitId,
+          );
+          const quantity = productValue.value * detail.quantity;
+          const increment = this.stock.increment(detail.productId, quantity);
+          if (increment) {
+            await this.repository.delete({ id });
+          }
+        });
+      }
       return {
         message: 'Delete order successfully',
         error: null,
@@ -171,7 +230,19 @@ export class OrderService {
           status: HttpStatus.NOT_FOUND,
         };
       }
-      await this.repository.softDelete({ id });
+      if (order) {
+        order.orderDetails.map(async (detail) => {
+          const productValue = await this.product.findValueProductByUnit(
+            detail.productId,
+            detail.unitId,
+          );
+          const quantity = productValue.value * detail.quantity;
+          const increment = this.stock.increment(detail.productId, quantity);
+          if (increment) {
+            await this.repository.softDelete({ id });
+          }
+        });
+      }
       return {
         message: 'Soft delete order successfully',
         error: null,
@@ -199,7 +270,19 @@ export class OrderService {
           status: HttpStatus.NOT_FOUND,
         };
       }
-      await this.repository.recover({ id });
+      if (order) {
+        order.orderDetails.map(async (detail) => {
+          const productValue = await this.product.findValueProductByUnit(
+            detail.productId,
+            detail.unitId,
+          );
+          const quantity = productValue.value * detail.quantity;
+          const increment = this.stock.decrement(detail.productId, quantity);
+          if (increment) {
+            await this.repository.recover({ id });
+          }
+        });
+      }
       return {
         message: 'Restore order successfully',
         error: null,
